# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateMapDiscussionCommentRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
  handlers.CreateMapDiscussionRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
...
  handlers.EditMapDiscussionRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
...
  handlers.LoginResponse:
    properties:
      token:
        type: string
    type: object
  handlers.MapDiscussion:
    properties:
      comments:
        items:
          $ref: '#/definitions/handlers.MapDiscussionComment'
        type: array
      content:
        type: string
      created_at:
        description: Upvotes   int                        `json:"upvotes"`
        type: string
      creator:
        $ref: '#/definitions/models.UserShortWithAvatar'
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  handlers.MapDiscussionComment:
    properties:
      comment:
        type: string
      date:
        type: string
      user:
        $ref: '#/definitions/models.UserShortWithAvatar'
    type: object
  handlers.MapDiscussionResponse:
    properties:
      discussion:
        $ref: '#/definitions/handlers.MapDiscussion'
    type: object
  handlers.MapDiscussionsResponse:
    properties:
      discussions:
        items:
          $ref: '#/definitions/handlers.MapDiscussion'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.UserShortWithAvatar:
    properties:
      avatar_link:
        type: string
      steam_id:
        type: string
      user_name:
        type: string
    type: object
host: lp.pektezol.dev
info:
  contact: {}
  description: Backend API endpoints for Least Portals Hub.
  license:
    name: GNU Affero General Public License, Version 3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  title: Least Portals Hub
  version: "1.0"
paths:
...
  /maps/{mapid}/discussions:
    get:
      description: Get map discussions with specified map id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MapDiscussionsResponse'
              type: object
      tags:
      - maps / discussions
    post:
      description: Create map discussion with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMapDiscussionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateMapDiscussionRequest'
              type: object
      tags:
      - maps / discussions
  /maps/{mapid}/discussions/{discussionid}:
    delete:
      description: Delete map discussion with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Discussion ID
        in: path
        name: discussionid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - maps / discussions
    get:
      description: Get map discussion with specified map and discussion id.
      parameters:
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Discussion ID
        in: path
        name: discussionid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.MapDiscussionResponse'
              type: object
      tags:
      - maps / discussions
    post:
      description: Create map discussion comment with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Discussion ID
        in: path
        name: discussionid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMapDiscussionCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.CreateMapDiscussionCommentRequest'
              type: object
      tags:
      - maps / discussions
    put:
      description: Edit map discussion with specified map id.
      parameters:
      - description: JWT Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Map ID
        in: path
        name: mapid
        required: true
        type: integer
      - description: Discussion ID
        in: path
        name: discussionid
        required: true
        type: integer
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EditMapDiscussionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EditMapDiscussionRequest'
              type: object
      tags:
      - maps / discussions
...
  /token:
    delete:
      description: Deletes the token cookie from the user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.LoginResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - auth
    get:
      description: Gets the token cookie value from the user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.LoginResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - auth
...
swagger: "2.0"